renamed imports from profile page
