+++ client/src/components/ui/index.ts
@@ -19,4 +19,4 @@ export { default as Dropdown } from './Dropdown';
 export { default as SelectDropDown } from './SelectDropDown';
 export { default as DialogTemplate } from './DialogTemplate';
 export { default as MultiSelectDropDown } from './MultiSelectDropDown';
-export { default as Profile } from './Profile';
+export { default as Profile } from '../Profile';
